+++ changelog.md
@@ -18,6 +18,7 @@ Codion Change Log
 - EntityType.resourceBundleName() now returns Optional.
 - AttributeDefinition.minimumValue() and maximumValue() now return Optional.
 - AttributeDefinition.format(), dateTimePattern() and dateTimeFormatter() now return Optional.
+- AttributeDefinition.description() now returns Optional.
 ### is.codion.framework.model
 - AbstractEntityEditModel.refresh() bug fixed, now handles the case when a primary key attribute has been modified.
 - AbstractEntityTableModel.refresh(keys) bug fixed, now includes the same attributes as the query model, when selecting.